+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- Validate `cilium.giantswarm.io/pod-cidr` annotation does not intersect with current CIDR's.
 - Validate `cilium.giantswarm.io/pod-cidr` annotation is present and valid while upgrading from v17 to v18.
 
 ## [4.1.0] - 2022-07-22